move files into Showings folder
