+++ src/App.tsx
@@ -3,7 +3,7 @@ import "./App.css"
 import Login from "./Login"
 import UserContext from "../hooks/UserContext.ts"
 import { User } from "../types.ts"
-import Showings from "./Showings.tsx"
+import Showings from "./Showings/Showings.tsx"
 
 function App() {
     const [user, setUser] = useState<User>({ role: "guest", username: "guest" })
