+++ src/Showings/ShowingCard.css
