+++ src/Showings/ShowingCard.tsx
@@ -7,8 +7,8 @@ function ShowingCard({showing}: {showing: any}) {
 
     return <button className="showing-card" onClick={()=>console.log(`${showing.name}`)}>
         <p>{showing.name}</p>
-        <p>{`${day}`}</p>
-        <p>{`${time}`}</p>
+        <p>{day}</p>
+        <p>{time}</p>
         <img src={`${showing.poster}`} alt={`Movie post for ${showing.film}`} className="poster"/>
     </button>
 }
